rpc, utils: move checkAndRaiseFDLimit to SetKlayConfig
